pyproject.toml
[tool.poetry]
name = "memo_1"
version = "0.1.0"
description = ""
authors = ["Sion Noh <shtldhs123@ajou.ac.kr>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
bs4 = "^0.0.1"
Flask = "^2.0.0"
pymongo = "^3.11.4"
PyJWT = "^2.1.0"
python-dotenv = "^0.17.1"
pytest-flask = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
